Eliminate unwanted entity
After moving model responsibilities from recommendations presenter to the recommendations manager, the initial recommendations model became just a proxy for the catalogue service. It made sense to move the latter's single responsibility to the manager, to remove the model and to make the manager the model.
