+++ Worm/Sources/Screens/Recommendations/RecommendationsPresenter.swift
@@ -30,7 +30,7 @@ protocol RecommendationsPresenter: ObservableObject {
 // MARK: -
 
 /// The default implementation of the Recommendations screen presenter.
-final class RecommendationsDefaultPresenter<Manager: RecommendationsManager>: RecommendationsPresenter {
+final class RecommendationsDefaultPresenter<Model: RecommendationsModel>: RecommendationsPresenter {
 
     // MARK: - Properties
 
@@ -41,13 +41,13 @@ final class RecommendationsDefaultPresenter<Manager: RecommendationsManager>: Re
 
     // MARK: Private properties
 
-    private let model: RecommendationsModel
-    private let recommendationsManager: Manager
+    private let model: Model
     private let updateQueue: DispatchQueue
     private lazy var cancellables = Set<AnyCancellable>()
 
     // MARK: - Initialization
 
+    // TODO: Update HeaderDoc.
     /**
      Creates a presenter object.
      - Parameters:
@@ -55,9 +55,8 @@ final class RecommendationsDefaultPresenter<Manager: RecommendationsManager>: Re
         - recommendationsManager: Object that owns logic of maintaing a list of recommedations.
         - updateQueue: Queue on which presentation data is passed to view.
      */
-    init(model: RecommendationsModel, recommendationsManager: Manager, updateQueue: DispatchQueue = .main) {
-        self.model = model
-        self.recommendationsManager = recommendationsManager
+    init(recommendationsManager: Model, updateQueue: DispatchQueue = .main) {
+        self.model = recommendationsManager
         self.updateQueue = updateQueue
     }
 
@@ -66,13 +65,12 @@ final class RecommendationsDefaultPresenter<Manager: RecommendationsManager>: Re
     // MARK: RecommendationsPresenter protocol methods
 
     func onViewAppear() {
-        bind(recommendationsManager: recommendationsManager)
-        updateFavoriteBooks()
+        bind(recommendationsManager: model)
     }
 
     // MARK: Private methods
 
-    private func bind(recommendationsManager: Manager) {
+    private func bind(recommendationsManager: Model) {
         recommendationsManager
             .objectWillChange
             .receive(on: updateQueue)
@@ -83,20 +81,6 @@ final class RecommendationsDefaultPresenter<Manager: RecommendationsManager>: Re
                 self?.recommendations = recommendationsManager.recommendations.map { $0.asViewModel(favorite: true) }
         }
         .store(in: &cancellables)
-    }
-
-    private func updateFavoriteBooks() {
-        model.favoriteBookIDs.forEach { addSimilarBooksToRecommendations(from: $0) }
-    }
-
-    private func addSimilarBooksToRecommendations(from bookID: String) {
-        model.getBook(by: bookID) { [weak self] in
-            self?.addSimilarBooksToRecommendations(from: $0?.similarBookIDs ?? [])
-        }
-    }
-
-    private func addSimilarBooksToRecommendations(from ids: [String]) {
-        ids.forEach { self.recommendationsManager.addRecommendation(id: $0) }
     }
 
 }
